Update README.md
The Quditto emulation platform has been added. It is an open-source project that allows the automatic deployment of distributed QKD networks on any Python-capable hardware. The nodes of the emulated network follow the ETSI 014 API, making the interaction between users and said nodes the same as between users and QKD hardware.